+++ test/test_quantized.py
@@ -6,25 +6,9 @@ import numpy as np
 import unittest
 from common_utils import TestCase, run_tests, skipIfNotRegistered
 
+
 def canonical(graph):
     return str(torch._C._jit_pass_canonicalize(graph))
-
-
-# Reference method for quantizing a tensor.
-def _fake_quantize_reference(X, scale, zero_point, num_bits):
-    quant_min, quant_max = 0, 2 ** num_bits - 1
-    res = (np.clip(np.round(X / scale) + zero_point, quant_min, quant_max) - zero_point) * scale
-    res = res.reshape(X.shape)
-    return res
-
-
-# Reference method for the gradient of the quantizer.
-def _fake_quantize_grad_reference(X, dY, scale, zero_point, num_bits):
-    Xq = np.round(X / scale) + zero_point
-    quant_min, quant_max = 0, 2 ** num_bits - 1
-    mask = np.logical_and(Xq >= quant_min, Xq <= quant_max)
-    res = dY[mask].reshape(dY.shape)
-    return res
 
 
 @skipIfNotRegistered("Relu_ENGINE_DNNLOWP",
@@ -79,44 +63,6 @@ class TestQuantizedRelu(unittest.TestCase):
         Y_hat_tensor = Y_hat[0].numpy()
 
         np.testing.assert_equal(Y_tensor, Y_hat_tensor)
-
-
-class TestFakeQuantize(unittest.TestCase):
-    """Tests the forward path of the fake quantization minMax args op."""
-    def test_forward(self):
-        fake_quantize_forward = torch.ops.quantized.fake_quantize_forward
-
-        scale = 3
-        zero_point = 2
-        num_bits = 8
-        X = np.random.rand(20, 20) * 125
-        Y = _fake_quantize_reference(X, scale, zero_point, num_bits)
-        X_torch = torch.from_numpy(X)
-        Y_prime = fake_quantize_forward(
-            X=X_torch, scale=scale, zero_point=zero_point, num_bits=num_bits,
-            quant_delay=0, iter=0)
-        tolerance = 1e-6
-        np.testing.assert_allclose(Y, Y_prime, rtol=tolerance, atol=tolerance)
-
-    """Tests the backward method. Note that this runs the reference quantization
-    and thus the errors might be originating there."""
-    def test_backward(self):
-        fake_quantize_backward = torch.ops.quantized.fake_quantize_backward
-
-        scale = 3
-        zero_point = 2
-        num_bits = 8
-        X = np.random.rand(20, 20) * 125
-        Y = _fake_quantize_reference(X, scale, zero_point, num_bits)
-        dY = Y - X  # Fake gradient
-        dX = _fake_quantize_grad_reference(X, dY, scale, zero_point, num_bits)
-        X_torch = torch.from_numpy(X)
-        dY_torch = torch.from_numpy(dY)
-        dX_prime = fake_quantize_backward(
-            X=X_torch, dY=dY_torch, scale=scale, zero_point=zero_point,
-            num_bits=num_bits, quant_delay=0, iter=0)
-        tolerance = 1e-6
-        np.testing.assert_allclose(dX, dX_prime, rtol=tolerance, atol=tolerance)
 
 
 if __name__ == '__main__':
